Added intellij-common-library and removed duplicated/same classes
