+++ build.gradle
@@ -9,7 +9,14 @@ plugins {
 downloadRobotServerPlugin.version = remoteRobotVersion
 
 repositories {
+    mavenLocal()
     mavenCentral()
+    maven {
+        url 'https://repository.jboss.org/nexus/content/repositories/snapshots'
+    }
+    maven {
+        url 'https://repository.jboss.org/nexus/content/groups/public'
+    }
     maven {
         url 'https://repository.jboss.org'
     }
@@ -57,6 +64,7 @@ sourceSets {
 }
 
 task integrationTest(type: Test) {
+    useJUnitPlatform()
     description = 'Runs the integration tests.'
     group = 'verification'
     testClassesDirs = sourceSets.integrationTest.output.classesDirs
@@ -66,24 +74,15 @@ task integrationTest(type: Test) {
 }
 
 dependencies {
-    implementation 'org.junit.jupiter:junit-jupiter:5.7.0'
     compile 'com.redhat.devtools.intellij:intellij-common:1.1.0'
     compile 'com.squareup.okhttp3:okhttp:3.12.12'
     compile group: 'org.eclipse.lsp4j', name: 'org.eclipse.lsp4j.jsonrpc', version: '0.9.0'
     compile 'org.jboss.tools.rsp.api.bundles:org.jboss.tools.rsp.api:0.22.9.Final@jar'
-    integrationTestImplementation(
-            "com.intellij.remoterobot:remote-robot:${remoteRobotVersion}",
-            "com.intellij.remoterobot:remote-fixtures:${fixturesVersion}"
-    )
+    compile 'com.redhat.devtools.intellij:intellij-common-ui-test-library:0.0.2'
 }
 
 runIdeForUiTests {
-    systemProperty "robot-server.port", "8082" // default port 8080
-    systemProperties['com.redhat.devtools.intellij.telemetry.mode'] = 'debug'
-}
-
-integrationTest {
-    useJUnitPlatform()
+    systemProperty "robot-server.port", System.getProperty("robot-server.port")
 }
 
 group 'com.redhat.devtools.intellij'
