--- src/it/java/com/redhat/devtools/intellij/rsp/dialogs/WelcomeFrameDialogFixture.java
@@ -1,47 +1,0 @@
-/*******************************************************************************
- * Copyright (c) 2020 Red Hat, Inc.
- * Distributed under license by Red Hat, Inc. All rights reserved.
- * This program is made available under the terms of the
- * Eclipse Public License v2.0 which accompanies this distribution,
- * and is available at https://www.eclipse.org/legal/epl-v20.html
- *
- * Contributors:
- * Red Hat, Inc. - initial API and implementation
- ******************************************************************************/
-package com.redhat.devtools.intellij.rsp.dialogs;
-
-import com.intellij.remoterobot.RemoteRobot;
-import com.intellij.remoterobot.data.RemoteComponent;
-import com.intellij.remoterobot.fixtures.ComponentFixture;
-import com.intellij.remoterobot.fixtures.ContainerFixture;
-import com.intellij.remoterobot.fixtures.DefaultXpath;
-import com.intellij.remoterobot.fixtures.FixtureName;
-import com.redhat.devtools.intellij.rsp.other.ActionLinkFixture;
-import org.jetbrains.annotations.NotNull;
-
-import static com.intellij.remoterobot.search.locators.Locators.byXpath;
-
-/**
- * Welcome to IntelliJ IDEA dialog fixture
- *
- * @author zcervink@redhat.com
- */
-@DefaultXpath(by = "FlatWelcomeFrame type", xpath = "//div[@class='FlatWelcomeFrame']")
-@FixtureName(name = "Welcome To IntelliJ IDEA Dialog")
-public class WelcomeFrameDialogFixture extends ContainerFixture {
-    public WelcomeFrameDialogFixture(@NotNull RemoteRobot remoteRobot, @NotNull RemoteComponent remoteComponent) {
-        super(remoteRobot, remoteComponent);
-    }
-
-    public ActionLinkFixture createNewProjectLink() {
-        return find(ActionLinkFixture.class, byXpath("//div[@text='New Project' and @class='ActionLink']"));
-    }
-
-    public ComponentFixture importProjectLink() {
-        return find(ComponentFixture.class, byXpath("//div[@text='Open or Import' and @class='ActionLink']"));
-    }
-
-    public ComponentFixture ideErrorsIcon() {
-        return find(ComponentFixture.class, byXpath("//div[@class='IdeErrorsIcon']"));
-    }
-}
