+++ src/it/java/com/redhat/devtools/intellij/rsp/menus/RightClickMenu.java
@@ -1,5 +1,5 @@
 /*******************************************************************************
- * Copyright (c) 2020 Red Hat, Inc.
+ * Copyright (c) 2021 Red Hat, Inc.
  * Distributed under license by Red Hat, Inc. All rights reserved.
  * This program is made available under the terms of the
  * Eclipse Public License v2.0 which accompanies this distribution,
@@ -8,7 +8,7 @@
  * Contributors:
  * Red Hat, Inc. - initial API and implementation
  ******************************************************************************/
-package com.redhat.devtools.intellij.rsp.dialogs;
+package com.redhat.devtools.intellij.rsp.menus;
 
 import com.intellij.remoterobot.RemoteRobot;
 import com.intellij.remoterobot.data.RemoteComponent;
@@ -17,15 +17,20 @@ import com.intellij.remoterobot.fixtures.DefaultXpath;
 import com.intellij.remoterobot.fixtures.FixtureName;
 import org.jetbrains.annotations.NotNull;
 
+import static com.intellij.remoterobot.search.locators.Locators.byXpath;
+
 /**
- * Project Structure dialog fixture
- *
  * @author olkornii@redhat.com
  */
-@DefaultXpath(by = "MyDialog type", xpath = "//div[@accessiblename='Project Structure' and @class='MyDialog']")
-@FixtureName(name = "Tip Of The Day Dialog")
-public class ProjectStructureDialogFixture extends CommonContainerFixture {
-    public ProjectStructureDialogFixture(@NotNull RemoteRobot remoteRobot, @NotNull RemoteComponent remoteComponent) {
+@DefaultXpath(by = "MyMenu type", xpath = "//div[@class='MyMenu']")
+@FixtureName(name = "My Menu")
+public class RightClickMenu extends CommonContainerFixture {
+    public RightClickMenu(@NotNull RemoteRobot remoteRobot, @NotNull RemoteComponent remoteComponent) {
         super(remoteRobot, remoteComponent);
     }
+
+    public void select(String selectedItem){
+        find(RightClickMenu.class, byXpath("//div[@accessiblename='" + selectedItem + "' and @class='ActionMenuItem']")).click();
+    }
 }
+
