--- src/it/java/com/redhat/devtools/intellij/rsp/other/ActionLinkFixture.java
@@ -1,35 +1,0 @@
-/*******************************************************************************
- * Copyright (c) 2020 Red Hat, Inc.
- * Distributed under license by Red Hat, Inc. All rights reserved.
- * This program is made available under the terms of the
- * Eclipse Public License v2.0 which accompanies this distribution,
- * and is available at https://www.eclipse.org/legal/epl-v20.html
- *
- * Contributors:
- * Red Hat, Inc. - initial API and implementation
- ******************************************************************************/
-package com.redhat.devtools.intellij.rsp.other;
-
-import com.intellij.remoterobot.RemoteRobot;
-import com.intellij.remoterobot.data.RemoteComponent;
-import com.intellij.remoterobot.fixtures.ComponentFixture;
-import com.intellij.remoterobot.fixtures.FixtureName;
-import org.jetbrains.annotations.NotNull;
-
-/**
- * Action link fixture
- *
- * @author zcervink@redhat.com
- */
-@FixtureName(name = "Action Link")
-public class ActionLinkFixture extends ComponentFixture {
-    public ActionLinkFixture(@NotNull RemoteRobot remoteRobot, @NotNull RemoteComponent remoteComponent) {
-        super(remoteRobot, remoteComponent);
-    }
-
-    public void click() {
-        runJs("const offset = component.getHeight()/2;\n" +
-                "robot.click(component, new Point(offset, offset), MouseButton.LEFT_BUTTON, 1);"
-        );
-    }
-}
